Fix alignment and represent pointers in alignment units
Pointers are now represented in "alignment units", since the typed array views
we use only allow aligned access. That is, an *int will now always be in
multiples of 4, so we store the pointer as the index into the Int32Array
instead of storing the byte address, or the index into the Int8Array. Put
another way, the pointer is divided by the size of its base ahead of time.

This makes emitted code cleaner to read, gets rid of shifts on memory access,
but adds shifts on pointer comparison and pointer arithmetic on pointers to
structs.